# add torch sdpa to `TransformerEncoder`